feat: add friends.module.ts, friend.model.ts, update user.model to add belong-to-many relation with friends table, move app.module.ts from root src to app folder and add the env variable to app.listen params in main.ts
