+++ src/features/app/app.module.ts
@@ -1,16 +1,18 @@
 import { Module } from '@nestjs/common';
 import { AppController } from 'src/features/app/app.controller';
-import { AppService } from './features/app/app.service';
+import { AppService } from './app.service';
 import { SequelizeModule } from '@nestjs/sequelize';
 import { User } from 'src/features/users/user.model';
 import { UsersModule } from 'src/features/users/users.module';
 import { AuthModule } from 'src/features/auth/auth.module';
 import { MessagesGatewayModule } from 'src/features/messages-websocket/messages.gateway.module';
-import { ServersModule } from './features/servers/servers.module';
-import { AuthGuard } from './features/auth/auth.guard';
-import { ServerUser } from './features/server-user/server-user.model';
-import { Server } from './features/servers/server.model';
-import * as config from '../config/config.json';
+import { ServersModule } from '../servers/servers.module';
+import { AuthGuard } from '../auth/auth.guard';
+import { ServerUser } from '../server-user/server-user.model';
+import { Server } from '../servers/server.model';
+import * as config from '../../../config/config.json';
+import { FriendsModule } from '../friends/friends.module';
+import { Friends } from '../friends/models/friend.model';
 
 const DbDevConfig = config.development;
 @Module({
@@ -18,7 +20,7 @@ const DbDevConfig = config.development;
     SequelizeModule.forRoot({
       ...DbDevConfig,
       dialect: 'postgres',
-      models: [User, Server, ServerUser],
+      models: [User, Server, ServerUser, Friends],
       autoLoadModels: true,
       synchronize: true,
     }),
@@ -26,6 +28,7 @@ const DbDevConfig = config.development;
     AuthModule,
     MessagesGatewayModule,
     ServersModule,
+    FriendsModule,
   ],
   controllers: [AppController],
   providers: [
