+++ src/features/users/user.model.ts
@@ -1,6 +1,7 @@
 import { Column, Model, Table, BelongsToMany } from 'sequelize-typescript';
 import { Server } from '../servers/server.model';
 import { ServerUser } from 'src/features/server-user/server-user.model';
+import { Friends } from '../friends/models/friend.model';
 
 @Table({ tableName: 'users' })
 export class User extends Model {
@@ -30,4 +31,7 @@ export class User extends Model {
 
   @BelongsToMany(() => Server, () => ServerUser)
   servers: Server[];
+
+  @BelongsToMany(() => User, () => Friends)
+  friends: Friends[];
 }
